pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "woodward"
version = "2.0.0"
description = "Woodward Core v2 — Investigative journalism verification and publication platform"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2",
    "pydantic-settings>=2",
    "fastapi",
    "uvicorn[standard]",
    "aiosqlite",
    "pandas",
    "pyyaml",
    "openai",
    "anthropic",
    "google-generativeai",
    "lancedb",
    "pdfplumber",
    "python-dotenv",
    "rich",
    "typer[all]",
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
woodward = "src.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
